Move language configuration functions to typescript
